eliminate a lot of warnings

+++ fast_kmer_filter.cpp
@@ -102,7 +102,7 @@ int main(int argc, char **argv) {
     		#pragma omp parallel for schedule(dynamic, 1)
 	    	for (int i = 0; i < batch_size; i++) {
 	    	// vector<ProtKmerGenerator> kmer_gens;
-			   	if (sequence_storage[i].sequence_.size() >= kmer_size) {
+			   	if ((int)sequence_storage[i].sequence_.size() >= kmer_size) {
 			   		ProcessSequenceMulti(sequence_storage[i].sequence_, sequence_storage[i].name_,sequence_storage[i].comment_, kmerSet, kmer_size);
 			   		RevComp(sequence_storage[i].sequence_);
 			   		ProcessSequenceMulti(sequence_storage[i].sequence_, sequence_storage[i].name_,sequence_storage[i].comment_, kmerSet, kmer_size);
@@ -118,7 +118,7 @@ int main(int argc, char **argv) {
     	// timer.start();
     	#pragma omp parallel for schedule(dynamic, 1)
     	for (int i = 0; i < count; i++) {
-		   	if (sequence_storage[i].sequence_.size() >= kmer_size) {
+		   	if ((int)sequence_storage[i].sequence_.size() >= kmer_size) {
 		   		ProcessSequenceMulti(sequence_storage[i].sequence_, sequence_storage[i].name_,sequence_storage[i].comment_, kmerSet, kmer_size);
 		   		RevComp(sequence_storage[i].sequence_);
 		   		ProcessSequenceMulti(sequence_storage[i].sequence_, sequence_storage[i].name_,sequence_storage[i].comment_, kmerSet, kmer_size);
